Reworked JSON to fix compatibility, added dynamic node name
Reworked JSON to fix a compatibility issue with CentOS6/ ElasticSearch 2.4+ (Specifically around CPU utilization and Memory utilization). Also added heap percent utilization monitoring.

Used socket to build a dynamic URL and Node value finder using the LAN IP address of the host to construct the URL and the local hostname as the node name. This correlates to Elastic industry best practice in regards to naming conventions as well as improves ease of deployment for large scale elastic instances.