Replaced faulty error handling
Error handling caused all nodes to return in an error state due to improper use of if else loop. Replaced with new variable to tell if if NODE == nodename was actually run. If it was not run (thus keeping nodecount at 0) it will error out with Node Not Found In Cluster.